requirements.txt
aspose_cells==24.9.0
JPype1==1.5.0
JPype1==1.5.0
openai==1.47.1
pandas==2.2.3
pdf2image==1.17.0
Pillow==10.4.0
